move to pvarray, address review comments
