+++ docs/sphinx/source/reference/pv_modeling/system_models.rst
@@ -54,4 +54,4 @@ PVGIS model
 .. autosummary::
    :toctree: ../generated/
 
-    pvsystem.huld
+    pvarray.huld
